refactor: 공통 UI 패키지 기반 구축
