+++ apps/admin/package.json
@@ -17,23 +17,18 @@
     "check": "biome check"
   },
   "dependencies": {
-    "@radix-ui/react-label": "^2.1.2",
-    "@radix-ui/react-slot": "^1.2.4",
-    "@radix-ui/react-tabs": "^1.1.13",
     "@solid-connect/api-schema": "workspace:^",
+    "@solid-connect/ui": "workspace:^",
     "@stomp/stompjs": "^7.1.1",
     "@tailwindcss/vite": "^4.0.6",
     "@tanstack/react-query": "^5.84.1",
     "axios": "^1.6.7",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
     "lucide-react": "^0.561.0",
     "react": "^19.2.6",
     "react-dom": "^19.2.6",
     "sockjs-client": "^1.6.1",
     "sonner": "^2.0.7",
-    "tailwind-merge": "^3.0.2",
     "tailwindcss": "^4.0.6"
   },
   "devDependencies": {
