+++ apps/admin/src/styles.css
@@ -1,4 +1,5 @@
 @import "tailwindcss";
+@source "../../../packages/ui/src";
 
 body {
   @apply m-0;
